Fixed some bugs + refactor

+++ src/main/scala/actions/Checkout.scala
@@ -2,12 +2,15 @@ package actions
 
 import java.io.File
 
+import objects.Branch
 import utils.IOManager
 
 object Checkout {
-
   def checkout(branch_name: String) = {
-    IOManager.writeFile(s"${IOManager.getRepoDirPath().get}${File.separator}HEAD", s"refs/heads/${branch_name}")
-    println(s"Switched to $branch_name")
+    if(Branch.getAllExistingBranches().contains(new Branch(branch_name))){
+      IOManager.writeFile(s"${IOManager.getRepoDirPath().get}${File.separator}HEAD", s"refs/heads/${branch_name}")
+      println(s"Switched to $branch_name")
+    } else println(s"Branch $branch_name doesn't exist.")
+
   }
 }
